merge

+++ ml4cvd/hyperparameters.py
@@ -336,7 +336,7 @@ def set_args_from_x(args, x):
         if k in x:
             logging.info(f"k is {k} and x[k] is {x[k]} args dict is:{args.__dict__[k]}")
             if k in ['conv_x', 'conv_y', 'conv_z']:
-                args.__dict__[k] = [float(x[k])]
+                args.__dict__[k] = [int(x[k])]
             elif isinstance(args.__dict__[k], int):
                 args.__dict__[k] = int(x[k])
             elif isinstance(args.__dict__[k], float):
